Fix permission issues in CI
Previously the override file was here, I removed the line in a previous
commit since the override file doesn't exist.

That means prior to this patch the js and css services still had the
volume in place and since CI runs with a different UID/GID, boom errors.

This should fix it.