Fix incorrect asset digest being used in prod
If multiple sprockets files are found, Rails doesn't pick the latest one
based on the modified time of the file.

Instead it picks the first one it finds. This could (and will) lead to
Rails loading old versions of your assets since it's not using the latest
manifest file.

This patch will delete all of the sprockets files except the last built
one and it does it in a way where there will never be a blip of time
there's not at least one sprockets file available to serve.