Make manifest file checker more robust
Find was throwing an error if no existing manifests were available. Now
it works with or without there being existing manifest files.

Basically now it works for your first deploy and all future deploys.

I'm not terribly happy with how many shellcheck ignores were needed but
we're searching for a pattern. If anyone wants to open a PR with an
alternative implementation without needing ignores go for it.